requirements.txt
gensim
flask
pymongo
pyYaml